+++ frontend/src/App.jsx
@@ -9,7 +9,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap-icons/font/bootstrap-icons.min.css";
 import "bootstrap/js/dist/dropdown";
 import "bootstrap/js/dist/modal";
-import "easymde/dist/easymde.min.css";
+import "font-awesome/css/font-awesome.min.css"; // For EasyMDE
 import './App.css';
 
 function App() {